% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNewlineToLongNames.R
\name{addNewlineToLongNames}
\alias{addNewlineToLongNames}
\title{Add Newline to Long Names in a Data Frame Column}
\usage{
addNewlineToLongNames(data, column_name, max_char = 30)
}
\arguments{
\item{data}{A dataframe containing the column to be modified.}

\item{column_name}{The name of the column within \code{data} that contains the variable names to be adjusted.}

\item{max_char}{The maximum number of characters allowed per line before inserting a newline.
Defaults to 30 characters.}
}
\value{
The modified dataframe with newline characters added to the specified column's entries.
}
\description{
This function takes a dataframe and a column name, then modifies the contents of that column
by inserting newline characters into long variable names. The aim is to ensure that no line
exceeds a specified maximum character length, improving readability in plots or tables where
space is limited.
}
\examples{
df <- data.frame(variable_name = c("This is a very long variable name that needs to be split",
                                   "Short name",
                                   "Another extremely long variable name that requires splitting"))
df <- addNewlineToLongNames(df, "variable_name")
print(df)
}
