% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorizeCelDataIntoBatches.R
\name{categorizeCelDataIntoBatches}
\alias{categorizeCelDataIntoBatches}
\title{Categorize CEL Files into Processing Batches Based on Scan Dates}
\usage{
categorizeCelDataIntoBatches(affyData)
}
\arguments{
\item{affyData}{An \code{AffyBatch} object.}
}
\value{
A data frame with columns 'FileName', "OriginalDateTime", 'FormattedDateTime',
and 'Batch', listing each CEL file, its scan date and time, and its assigned processing
batch number.
}
\description{
Takes an \code{AffyBatch} object and categorizes CEL files into processing batches based on the
precise date and time they were run, attempting to accommodate variable date formats.
}
\examples{
# Assuming `affyData` is your AffyBatch object:
# batches <- categorizeCelDataIntoBatches(affyData)
}
