% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArmsFromMetaDataGSE145513.R
\name{createArmsFromMetaDataGSE145513}
\alias{createArmsFromMetaDataGSE145513}
\title{Create ARM Column from Metadata in ExpressionSet}
\usage{
createArmsFromMetaDataGSE145513(eset)
}
\arguments{
\item{eset}{An \code{ExpressionSet} object containing metadata to be processed.}
}
\value{
An updated \code{ExpressionSet} object with a new \code{ARM} column in the metadata.
}
\description{
Takes an \code{ExpressionSet} object and creates a new metadata column \code{ARM} by combining and cleaning
information from \code{characteristics_ch1.1}, \code{characteristics_ch1.2}, and \code{characteristics_ch1.3}.
}
\examples{
# Assuming `eset` is your ExpressionSet object:
# esetUpdated <- createArmsFromMetaDataGSE145513(eset)
}
