% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEnrichmentPlots.R
\name{createEnrichmentPlots}
\alias{createEnrichmentPlots}
\title{Create Enrichment Plots for Top and Bottom Ranked Pathways}
\usage{
createEnrichmentPlots(fgseaRes, topTableData)
}
\arguments{
\item{fgseaRes}{Results from fgsea analysis.}

\item{topTableData}{A dataframe containing differential expression analysis results.
Must include columns for log fold change (\code{logFC}) and adjusted p-value (\code{adj.P.Val}).}
}
\description{
Generates enrichment plots for the top 10 pathways with the highest and lowest NES
from the results of a GSEA analysis performed using the fgsea package.
}
\examples{
# Assuming 'fgseaRes' contains the results from your fgsea analysis:
# createEnrichmentPlots(fgseaRes, topTableData)
}
