% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTilePlotCellTypeAnovaResults.R
\name{createTilePlotCellTypeAnovaResults}
\alias{createTilePlotCellTypeAnovaResults}
\title{Create Tile Plot for ANOVA Results}
\usage{
createTilePlotCellTypeAnovaResults(finalResults)
}
\arguments{
\item{finalResults}{Dataframe containing ANOVA results, including columns for ARM1, ARM2,
CellType, and neg.log.10.p.adj, which is the negative logarithm (base 10) of the
adjusted p-value. This dataframe is expected to be the output from \code{analyzeCellTypeANOVA}.}
}
\value{
A ggplot object displaying a tile plot, which can be further customized or directly
rendered in R graphical devices.
}
\description{
Visualizes the significance of ANOVA results between different arms for various cell types
using a tile plot. Each tile's color intensity represents the negative log (base 10) of
the adjusted p-value, highlighting the significance of differences observed.
}
\examples{
# Assuming 'finalResults' contains the output from `analyzeCellTypeANOVA`:
plot <- createTilePlotCellTypeAnovaResults(finalResults)
print(plot)

}
