% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTsnePlot.R
\name{createTsnePlot}
\alias{createTsnePlot}
\title{Produce t-SNE plots from ExpressionSet object}
\usage{
createTsnePlot(eset, phenotypeCols, perplexity = 30)
}
\arguments{
\item{eset}{An ExpressionSet object containing expression data in the assayData slot
and phenotype data in the phenoData slot.}

\item{phenotypeCols}{A character vector specifying which columns from the phenotype
data to use for color-coding the t-SNE plots.}

\item{perplexity}{Perplexity parameter for the t-SNE algorithm (default is 30).}
}
\value{
A ggplot object displaying a faceted plot of t-SNE visualizations, each facet
corresponding to a different phenotype variable for coloring.
}
\description{
This function creates t-SNE plots for visualizing high-dimensional expression data
contained in an ExpressionSet object, color-coded by specified phenotype columns.
Each phenotype column generates a separate plot, and the function returns a faceted
plot combining these individual plots.
}
\examples{
# eset is your ExpressionSet object with expression and phenotype data
# phenotypeCols might include columns like "Treatment" or "CellType"
tsneFacetedPlot <- createTsnePlot(eset, c("Treatment", "CellType"))

}
