% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineGenderWithMassiR.R
\name{determineGenderWithMassiR}
\alias{determineGenderWithMassiR}
\title{Determine Gender Using massiR Package}
\usage{
determineGenderWithMassiR(eset)
}
\arguments{
\item{eset}{An ExpressionSet object containing expression data and phenotype data.}
}
\value{
An ExpressionSet object with an added column in phenoData called PredictedGender.
}
\description{
Determine Gender Using massiR Package
}
\examples{
# Assuming 'eset' is your ExpressionSet object:
esetWithGender <- determineGenderWithMassiR(eset)
}
