% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineGenderWithMassiRFromRnaSeq.R
\name{determineGenderWithMassiRFromRnaSeq}
\alias{determineGenderWithMassiRFromRnaSeq}
\title{Determine Gender Using the massiR Package}
\usage{
determineGenderWithMassiRFromRnaSeq(eset)
}
\arguments{
\item{eset}{An ExpressionSet object containing expression data and phenotype data.}
}
\value{
An updated ExpressionSet object with an added column in phenoData named
PredictedGender, which contains the gender prediction for each sample.
}
\description{
This function predicts the gender of samples in an ExpressionSet object using
Y chromosome probes. It accounts for different platforms by mapping annotations
and assumes RNA-seq if the platform is not recognized.
}
\examples{
# Assuming 'eset' is an ExpressionSet object with expression and phenotype data:
esetWithGender <- determineGenderWithMassiR(eset)
}
