% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCellTypeProportions.R
\name{estimateCellTypeProportions}
\alias{estimateCellTypeProportions}
\title{Estimate Cell Type Proportions Using ImSig}
\usage{
estimateCellTypeProportions(eset, correlationThreshold = 0.7)
}
\arguments{
\item{eset}{An ExpressionSet object with expression data.}

\item{correlationThreshold}{The correlation threshold (r) for feature selection in ImSig.}
}
\value{
A dataframe with cell type proportions per sample.
}
\description{
Estimate Cell Type Proportions Using ImSig
}
\examples{
# Assuming 'eset' is your ExpressionSet object:
cellProportions <- estimateCellTypeProportions(eset, correlationThreshold = 0.7)
}
