% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterLowExpressedGenesRnaSeq.R
\name{filterLowExpressedGenesRnaSeq}
\alias{filterLowExpressedGenesRnaSeq}
\title{Filter Low Expressed Genes from an ExpressionSet}
\usage{
filterLowExpressedGenesRnaSeq(eset, minSamples = 2, countCutoff = 10)
}
\arguments{
\item{eset}{An ExpressionSet object containing RNA-seq count values.}

\item{minSamples}{The minimum number of samples in which the gene must have
counts greater than the cutoff value to be retained. Default is 2.}

\item{countCutoff}{The count value above which counts are considered
significant for the purpose of filtering. Default is 10.}
}
\value{
A new ExpressionSet object with filtered expression data and
corresponding featureData.
}
\description{
This function filters out genes from an ExpressionSet object based on
RNA-seq count values. It retains only genes with a specified minimum number
of samples having counts greater than a specified cutoff.
}
\examples{
# Assuming 'eset' is your ExpressionSet object with RNA-seq count data:
filteredEset <- filterLowExpressedGenesRnaSeq(eset, minSamples = 2, countCutoff = 10)

}
