% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMetadataToPCs.R
\name{matchMetadataToPCs}
\alias{matchMetadataToPCs}
\title{Match Metadata to Principal Components}
\usage{
matchMetadataToPCs(eset, metadataCols)
}
\arguments{
\item{eset}{An ExpressionSet object containing expression data and metadata.}

\item{metadataCols}{A character vector specifying the column names in the metadata to be analyzed.}
}
\value{
A dataframe listing the principal component ID, the percent of variation explained by the PC, the column name of the metadata, and the confidence that this covariate is explained by the principal component.
}
\description{
This function matches the metadata (covariates) in an ExpressionSet object to the principal components of expression data. It uses the degCovariates function from the DEGreport R library.
}
\examples{
# Assuming 'eset' is your ExpressionSet:
result <- matchMetadataToPCs(eset, metadataCols = c("covariate1", "covariate2"))
}
