% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performComBatBatchCorrection.R
\name{performComBatBatchCorrection}
\alias{performComBatBatchCorrection}
\title{Perform ComBat Batch Correction}
\usage{
performComBatBatchCorrection(affyData, batchInfo)
}
\arguments{
\item{affyData}{An \code{AffyBatch} object.}

\item{batchInfo}{A data frame with batch information for each sample, typically
produced by the \code{categorizeCelDataIntoBatches} function. Must contain columns
'FileName' and 'Batch'.}
}
\value{
A matrix of batch-corrected expression data.
}
\description{
Applies ComBat batch correction to an AffyBatch object's expression data,
using the batch information provided. This function relies on the \code{sva} package.
}
\examples{
# Assuming `affyData` is your AffyBatch object and `batchInfo` contains the batch data:
# correctedData <- performComBatBatchCorrection(affyData, batchInfo)
}
