% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performComBatBatchCorrectionPCA.R
\name{performComBatBatchCorrectionPCA}
\alias{performComBatBatchCorrectionPCA}
\title{Perform ComBat Batch Correction, Visualize PCA, and Return Corrected Data}
\usage{
performComBatBatchCorrectionPCA(
  eset,
  batch,
  mod = NULL,
  outputPath = "PCA_Batch_Correction.png"
)
}
\arguments{
\item{eset}{An ExpressionSet object used for analysis.}

\item{batch}{A vector specifying the batch for each sample.}

\item{mod}{A model matrix specifying covariates of interest (other than batch).}

\item{outputPath}{Optional path for saving the PCA plot.}
}
\value{
A matrix of batch-corrected expression data.
}
\description{
Perform ComBat Batch Correction, Visualize PCA, and Return Corrected Data
}
\examples{
# eset is your ExpressionSet, batch is your batch vector
correctedData <- performComBatBatchCorrectionPCA(eset, batch, outputPath="PCA_Batch_Correction.png")
}
