% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performGSEA.R
\name{performGSEA}
\alias{performGSEA}
\title{Perform Pre-Ranked GSEA Using fgsea with MSigDB Category Selection}
\usage{
performGSEA(
  fgseaRes,
  rankingMethod,
  msigdbCategory,
  minSize = 10,
  maxSize = 500,
  nperm = 10000
)
}
\arguments{
\item{rankingMethod}{Method to rank genes: 't_stat', 'fc', or 'signed_p'.}

\item{msigdbCategory}{Category of MSigDB gene sets to use, e.g., 'H' for Hallmark.}

\item{minSize}{Minimum size of the gene set to be tested.}

\item{maxSize}{Maximum size of the gene set to be tested.}

\item{nperm}{Number of permutations for estimating p-values.}

\item{topTableData}{Dataframe from topTable from limma DEG analysis that contains p-value, fold change, and t-stat info}
}
\value{
A list containing the fgsea results for the specified MSigDB category.
}
\description{
Performs GSEA on limma topTable results using fgsea, allowing users to specify the gene set category from MSigDB.
}
\examples{
# Assuming 'topTableData' is the result of :
# GSEAresults <- performGSEA(topTableData, rankingMethod = 't_stat', msigdbCategory = 'H',
#                            minSize = 10, maxSize = 500, nperm = 10000)
}
