% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performLeadingEdgeAnalysis.R
\name{performLeadingEdgeAnalysis}
\alias{performLeadingEdgeAnalysis}
\title{Perform Leading Edge Analysis from fgsea Results}
\usage{
performLeadingEdgeAnalysis(
  fgseaRes,
  direction = "up",
  maxGeneSize = 20,
  maxGeneSets = 25
)
}
\arguments{
\item{fgseaRes}{Data frame containing fgsea results.}

\item{direction}{Character specifying the direction of regulation ("up" or "down").}

\item{maxGeneSize}{Integer specifying the maximum number of leading edge gene to display. Default is 20.}

\item{maxGeneSets}{Integer specifying the maximum number of gene sets to display. Default is 25.}
}
\description{
Visualizes the most frequent genes in the leading edges of upregulated or downregulated gene sets from fgsea results using a tile plot.
}
\examples{
# Assuming 'fgseaRes' is your dataframe containing fgsea results:
performLeadingEdgeAnalysis(fgseaRes, direction = "up", maxGeneSize = 20, maxGeneSets = 25)
}
