% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTopGenesBoxplot.R
\name{plotTopGenesBoxplot}
\alias{plotTopGenesBoxplot}
\title{Plot Top Up- and Down-Regulated Genes from Limma Analysis}
\usage{
plotTopGenesBoxplot(topTableData, eset, group1, group2)
}
\arguments{
\item{topTableData}{A dataframe output from Limma's topTable function, containing columns for logFC,
adjusted p-values (\code{adj.P.Val}), T-statistics (\code{t}), and gene symbols.}

\item{eset}{An \code{ExpressionSet} object used in the Limma analysis, containing the expression data and
phenotype data of samples.}

\item{group1}{A string specifying the first comparison group or condition in the Limma contrast.}

\item{group2}{A string specifying the second comparison group or condition in the Limma contrast.}
}
\value{
A list containing two ggplot objects: one for the top up-regulated genes and another for the top
down-regulated genes, each saved as a PNG file.
}
\description{
This function generates two sets of box plots for the top 20 up- and down-regulated genes based on differential
expression analysis results from Limma. It showcases log2 expression values across specified comparison
groups. Genes are categorized as up- or down-regulated based on their T-statistics, and plots are annotated
with adjusted p-values.
}
\examples{
topTableData <- limma::topTable(fit, coef="conditiontreated", number=Inf)
eset <- yourExpressionSet
plotTopGenesBoxplot(topTableData, eset, "Condition1", "Condition2")
}
