% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAffyCELFiles.R
\name{readAffyCELFiles}
\alias{readAffyCELFiles}
\title{Read Affymetrix CEL Files from Directory}
\usage{
readAffyCELFiles(dirName)
}
\arguments{
\item{dirName}{The path to the directory containing the CEL files.}
}
\value{
An \code{AffyBatch} object containing the raw data from the CEL files.
}
\description{
Reads all Affymetrix .CEL files located in a specified directory and returns an \code{AffyBatch} object
containing the raw data from these files. This function requires the \code{affy} package.
}
\examples{
# Assuming the CEL files are located in "/Users/briankweiner/R_code/TestData/cel_files/":
# affyData <- readAffyCELFiles("/Users/briankweiner/R_code/TestData/cel_files/")
}
