% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGenomeAnnotations.R
\name{readGenomeAnnotations}
\alias{readGenomeAnnotations}
\title{Read Genome Annotations from GFF File}
\usage{
readGenomeAnnotations(file_path, features = NULL)
}
\arguments{
\item{file_path}{The path to the GFF file.}

\item{features}{Optional vector of feature types to filter (e.g., "gene", "exon"). If NULL, all features are returned.}
}
\value{
A data frame with columns for sequence, source, feature, start, end, score, strand, phase, and attributes.
}
\description{
This function reads genomic annotations from a GFF (General Feature Format) file and returns a data frame with the annotations.
The data frame will include columns for sequence name, source, feature type, start position, end position, score, strand, phase, and attributes.
}
\examples{
gff_annotations <- readGenomeAnnotations("/Users/briankweiner/R_code/TestData/gff_files/GCF_000005845.2_ASM584v2_genomic.gff")
head(gff_annotations)
}
