% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGEO.R
\name{readGEO}
\alias{readGEO}
\title{Read GEO Series Matrix File and Return an ExpressionSet}
\usage{
readGEO(file_path)
}
\arguments{
\item{file_path}{The path to the GEO series matrix file.}
}
\value{
An \code{ExpressionSet} object containing the expression data and sample metadata.
}
\description{
This function reads a GEO series matrix file, extracting gene expression data along with sample metadata,
and constructs an ExpressionSet object. The ExpressionSet object contains the expression matrix and
phenotypic data for the samples.
}
\examples{
geo_eset <- readGEO("/Users/briankweiner/R_code/TestData/geo_files/GSE53552_series_matrix.txt")
Biobase::exprs(geo_eset) # To access the expression data
Biobase::pData(geo_eset) # To access the sample metadata
}
