% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVCF.R
\name{readVCF}
\alias{readVCF}
\title{Read Variants from a VCF File}
\usage{
readVCF(file_path, include_genotypes = FALSE)
}
\arguments{
\item{file_path}{The path to the VCF file.}

\item{include_genotypes}{A logical value indicating whether to include genotype information for each sample. Defaults to FALSE.}
}
\value{
A data frame with variant information, including columns for chromosome, position, ID, reference allele,
alternate allele(s), quality, filter, and INFO fields. If \code{include_genotypes} is TRUE, additional columns
for genotype information will be included.
}
\description{
This function reads variant call format (VCF) files and extracts the variant information into a data frame.
The resulting data frame includes columns for the chromosome, position, ID, reference allele, alternate allele(s),
quality, filter, and information (INFO) fields. Additional columns for genotype information may be included if present.
}
\examples{
vcf_data <- readVCF("/Users/briankweiner/R_code/TestData/vcf_files/00-All.vcf")
head(vcf_data)
}
