% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformExpressionData.R
\name{transformExpressionData}
\alias{transformExpressionData}
\title{Transform Expression Data in ExpressionSet}
\usage{
transformExpressionData(eset)
}
\arguments{
\item{eset}{An \code{ExpressionSet} object.}
}
\value{
A modified \code{ExpressionSet} object with the expression data transformed.
}
\description{
This function modifies the expression matrix within an ExpressionSet object by performing
the following steps:
\enumerate{
\item Replaces any values less than 0 with 0.
\item Adds 1 to every value in the matrix.
\item Applies a log2 transformation to all values.
The modified ExpressionSet object is returned.
}
}
\examples{
# Assuming `eset` is your ExpressionSet object:
# esetTransformed <- transformExpressionData(eset)
}
