% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateEsetWithUniqueGenes.R
\name{updateEsetWithUniqueGenes}
\alias{updateEsetWithUniqueGenes}
\title{Update ExpressionSet with Unique Gene Names Based on Highest Median Expression}
\usage{
updateEsetWithUniqueGenes(eset)
}
\arguments{
\item{eset}{An ExpressionSet object with potentially non-unique gene names.}
}
\value{
An updated ExpressionSet object with unique gene names and probes selected based on the highest median expression.
}
\description{
Updates an ExpressionSet object to ensure unique gene names in assayData$exprs,
choosing the probe with the highest median expression for genes with multiple mappings.
}
