% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateEsetWithUniqueGenesWithColumnSpecified.R
\name{updateEsetWithUniqueGenesWithColumnSpecified}
\alias{updateEsetWithUniqueGenesWithColumnSpecified}
\title{Update ExpressionSet with Unique Gene Names Based on Highest Median Expression}
\usage{
updateEsetWithUniqueGenesWithColumnSpecified(
  eset,
  geneSymbolColumn = "Gene Symbol"
)
}
\arguments{
\item{eset}{An ExpressionSet object with potentially non-unique gene names.}

\item{geneSymbolColumn}{The name of the column in featureData where gene symbols are stored.
Default is "Gene Symbol". Can be set to another column name such as "external_gene_name".}
}
\value{
An updated ExpressionSet object with unique gene names and probes selected based on
the highest median expression.
}
\description{
Updates an ExpressionSet object to ensure unique gene names in assayData$exprs,
choosing the probe with the highest median expression for genes with multiple mappings.
Allows specifying the column name where gene symbols are stored.
}
