% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeCellTypeProportions.R
\name{visualizeCellTypeProportions}
\alias{visualizeCellTypeProportions}
\title{Visualize Cell Type Proportions}
\usage{
visualizeCellTypeProportions(cellTypeDF)
}
\arguments{
\item{cellTypeDF}{A dataframe containing cell type proportions per sample,
expected to be the output of \code{estimateCellTypeProportions}.}
}
\value{
A ggplot object showing faceted bar graphs for each cell class.
}
\description{
Visualize Cell Type Proportions
}
\examples{
# Assuming 'cellTypeDF' is your dataframe from `estimateCellTypeProportions`:
visualizeCellTypeProportions(cellTypeDF)
}
